STYLE: Consolidate structs into ShapePopulationStruct.h
