--- src/axisColorStruct.h
@@ -1,15 +1,0 @@
-#ifndef AXISCOLORSTRUCT_H
-#define AXISCOLORSTRUCT_H
-
-
-struct axisColorStruct
-{
-public :
-    double XAxiscolor[3];
-    double YAxiscolor[3];
-    double ZAxiscolor[3];
-
-    bool complementaryColor;
-    bool sameColor;
-};
-#endif // AXISCOLORSTRUCT_H
