--- src/cameraConfigStruct.h
@@ -1,19 +1,0 @@
-#ifndef CAMERACONFIGSTRUCT_H
-#define CAMERACONFIGSTRUCT_H
-
-struct cameraConfigStruct
-{
-	double pos_x;
-	double pos_y;
-	double pos_z;
-	double foc_x;
-	double foc_y;
-	double foc_z;
-	double view_vx;
-	double view_vy;
-	double view_vz;
-    double scale;
-};
-
-#endif
-
