--- src/colorPointStruct.h
@@ -1,12 +1,0 @@
-#ifndef COLORPOINTSTRUCT_H
-#define COLORPOINTSTRUCT_H
-
-struct colorPointStruct
-{
-	double pos;
-	double r;
-	double g;
-	double b;
-};
-
-#endif
