--- src/magnitudStruct.h
@@ -1,10 +1,0 @@
-#ifndef MAGNITUDSTRUCT_H
-#define MAGNITUDSTRUCT_H
-
-struct magnitudStruct
-{
-    double max;
-    double min;
-};
-
-#endif // MAGNITUDSTRUCT_H
